refactor: CarsTest 생성
